requirements.txt
PyQt5==5.15.9
pyinstaller==5.13.0
Pillow==10.0.0 